Clear DicesGameScreen and add background image there.
Also, delete unused coin.png,
